+++ src/window.rs
@@ -19,8 +19,8 @@ pub async fn run<G, U, R>(
 ) -> Result<()>
 where
     G: 'static,
-    U: FnMut(&mut G, &Input, f64) + 'static,
-    R: FnMut(&mut G, &mut [u32], f64) + 'static,
+    U: FnMut(&mut G, &Input, f32) + 'static,
+    R: FnMut(&mut G, &mut [u32], f32) + 'static,
 {
     // Build the window builder with the event loop the user supplied
     let event_loop = EventLoop::new();
@@ -67,7 +67,7 @@ where
         move |g| update(&mut g.game.0, &g.game.2, 0.1),
         move |g| {
             let frame_time = g.last_frame_time();
-            render(&mut g.game.0, &mut buffer, frame_time);
+            render(&mut g.game.0, &mut buffer, frame_time as f32);
 
             // Blit draws the pixels in RGBA format, but the pixels crate expects BGRA, so convert it
             g.game
